+++ flowtype/helpers.py
@@ -1,5 +1,4 @@
 import os
-import re
 import json
 import subprocess
 from collections import namedtuple