+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 *This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog](http://keepachangelog.com/).*
 
 ## Unreleased
+### Added
+  * Add a spinner singleton to show a loading process in status bar.
+
 ### Fixed
   * Fix `check_contents` result not showing on Flow server initialization.